+++ src/modules/LessonContentText/index.ts
@@ -1,0 +1,3 @@
+import LessonContentText from "./LessonContentText";
+
+export default LessonContentText;
