+++ src/modules/VideoPlayer/index.ts
@@ -1,0 +1,3 @@
+import VideoPlayer from "./VideoPlayer";
+
+export default VideoPlayer;
